Move onboarding classes to separate package
