+++ src/main/java/seedu/contax/ui/onboarding/OnboardingStoryManager.java
@@ -1,4 +1,4 @@
-package seedu.contax.logic;
+package seedu.contax.ui.onboarding;
 
 import java.util.ArrayList;
 import java.util.HashSet;
